+++ data/src/main/java/com/krykun/data/datasource/local/MoviesLocalDataSource.kt
@@ -6,5 +6,5 @@ import kotlinx.coroutines.flow.Flow
 interface MoviesLocalDataSource {
     suspend fun insertMovie(movie: Movie, playlistId: Long): Long
     fun removeMovieFromPlaylist(movie: Movie, playlistId: Long)
-    fun isMovieAdded(movieId: Int, playlistId: Long): Boolean
+    fun isMovieAdded(movieId: Int, playlistId: Int): Boolean
 }
